requirements.txt
streamlit
PyGithub
requests
plotly
pandas
networkx
reportlab
matplotlib
kaleido
python-dotenv
PyJWT
bcrypt
